pull some functionality into src/open_gira
