+++ workflow/scripts/concat_and_sum_slices.py
@@ -16,7 +16,7 @@ import pandas as pd
 from tqdm import tqdm
 
 from transport.utils import NO_GEOM_ERROR_MSG
-from network_components import natural_sort
+from open_gira.utils import natural_sort
 
 
 if __name__ == "__main__":
